[schedulers/cron] 减少不必要的类型转换

+++ schedulers/cron/next_test.go
@@ -282,7 +282,7 @@ func TestBound_next(t *testing.T) {
 	type test struct {
 		// 输入
 		typ   int
-		curr  uint8
+		curr  int
 		list  uint64
 		carry bool
